feat: Enhance RzViewContext with additional properties
Added IUrlHelper and FormUrl to the RzViewContext class. The IUrlHelper provides methods to build URLs within an application, while the FormUrl property returns the current action method URL as a possible form callback URL. This can be overridden manually in any form handler method and used inside of form Razor Component views. Also added a strongly-typed data model for EditContext, which is essentially a shortcut for EditContext.Model that does typecasting.