feat: enhance Htmx configuration handling
Introduced a new dependency in `HtmxConfigHeadOutlet` to inject `RzViewContext`. This allows us to access the Htmx configuration directly from the view context. Also, added a condition to check if any specific configuration is set and update it accordingly.

In `HtmxContext`, added properties for HttpContext and IOptionsSnapshot of HtmxConfig. These changes allow us to get the required services and set up the default configuration during initialization. Added a method `SetConfiguration` that updates the current configuration based on provided name.

These changes improve how we handle configurations, making it more flexible and efficient.