feat: Add Htmx context to RzController classes
Added a new property to the `RzController` and `RzControllerWithViews` classes that exposes the Htmx context for the current request. This allows for more granular control over HTTP requests within these controllers. Also, updated the version of Rizzy in the project file from 1.0.7 to 1.0.8 and made minor formatting changes in HomeController.cs without altering functionality.