feat: Enable use of field identifiers in form context
This update enables the use of field identifiers in the form context. It also ensures that an 'id' attribute is added to attributes if it doesn't exist already. Minor changes include formatting adjustments and a modification to how .NET objects are serialized for use with Alpine.js x-data attribute. The 'id' attribute is no longer required when using RzInputText in views, simplifying its usage.