feat(swap-service): add event for content item updates
This commit introduces an event in the HtmxSwapService that gets triggered whenever content items are updated. This allows other components, such as HtmxSwapContent, to react to these changes and update their state accordingly. The version of the Rizzy project has been bumped up to 1.2.0 due to this new feature.

In addition, unnecessary code related to rendering swap content in RzRazorComponentResult has been removed as it's no longer needed with the new event-driven approach.

The Weather component in the demo application now also uses this new feature by adding a sidebar inside itself using the SwapService.