feat(swap-service): add content clearing functionality
This update introduces a new feature to the Swap Service. Now, it's possible to clear all content items from the service. This is achieved by adding a `Clear` method in both `IHtmxSwapService` interface and its implementation `HtmxSwapService`.

In addition, the `HtmxSwapContent` component has been refactored to store the rendered fragment during initialization and use it later when building the render tree. This change ensures that the content is only rendered once and then cleared from the swap service.

The `RzRazorComponentResult` class now checks if there's any available content in swap service after rendering a page. If so, it renders this additional content into response.

Lastly, version number of Rizzy project has been updated from 1.2.1 to 1.2.2.