feat: add PartialView method for RenderFragment
This commit introduces a new feature to the Rizzy framework. A `PartialView` method that accepts a `RenderFragment` parameter has been added to the `IRizzyService`, `RzController`, and `RzControllerWithViews`. This allows Razor components to be rendered without a layout from a RenderFragment.

In addition, the codebase has been updated to use C# 10's required properties in several places, replacing default values with required ones.

The version of the Rizzy project has also been bumped up from 2.0.8 to 2.0.9 as part of this change.

A new class, `FragmentComponent`, was created under Rizzy.Components namespace which takes care of building render tree if Fragment is not null.

Documentation for these changes have also been added in markdown files under docs directory.