feat: Introduced RzFormFieldMap and updated form components
Introduced a new class, RzFormFieldMap, to represent the FieldIdentifier to field name mapping for a given field. Updated form components (RzInputCheckbox and RzInputDate) to use this new class for managing field mappings.

Also, refactored the code in these components to replace usage of RzEditForm with EditForm and introduced usage of ViewContext. This change simplifies the management of field mappings by centralizing it within ViewContext.

Additionally, commented out unused classes (RzEditForm and RzFormContext) as part of code cleanup.