feat: Expanded nonce functionality
Expanded the functionality of the RizzyNonceProvider to support various types of nonces (script, style, font, image, connect) for use in Content Security Policies. This includes changes to method signatures and comments in IRizzyNonceProvider.cs and RizzyNonceProvider.cs. Also added a new enum NonceType.cs to define different types of nonces.