feat: Update project version and content handling
The project version has been updated. Additionally, the way static files are handled has been changed. Instead of using custom targets to copy these files during build and publish, they are now included as Content with the 'PreserveNewest' setting for the 'CopyToOutputDirectory' property. This simplifies the csproj file and ensures that only the newest versions of these files are used.